requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==3.2.13
django-cors-headers==3.13.0
django-environ==0.8.1
django-extensions==3.2.3
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==20.1.0
html5lib==1.1
idna==3.6
l18n==2021.3
mysqlclient==2.2.0
openpyxl==3.1.2
pillow==10.2.0
pillow-heif==0.14.0
PyJWT==2.4.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.8.0
urllib3==2.1.0
wagtail==5.2.3
webencodings==0.5.1
Willow==1.6.3
